Create Telephone Number Validator
The user may fill out the form field any way they choose as long as it has the format of a valid US number. Your job is to validate or reject the US phone number based on any combination of the formats provided above. The area code is required. If the country code is provided, you must confirm that the country code is 1. Return true if the string is a valid US phone number; otherwise, return false.